requirements.txt
fastapi==0.68.1
Rastreio-Correios==0.1.4
uvicorn==0.15.0
mangum==0.12.2